Move attributions manager
